Wrote TaskController

+++ sidewalk-webpage/app/formats/json/TaskSubmissionFormats.scala
@@ -8,9 +8,9 @@ object TaskSubmissionFormats {
   case class EnvironmentSubmission(browser: Option[String], browserVersion: Option[String], browserWidth: Option[Int], browserHeight: Option[Int],
                                   availWidth: Option[Int], availHeight: Option[Int], screenWidth: Option[Int], screenHeight: Option[Int], operatingSystem: Option[String])
 
-  case class InteractionSubmission(action: String, gsv_panorama_id: Option[String], lat: Option[Double], lng: Option[Double], heading: Option[Double], pitch: Option[Double], zoom: Option[Int], note: Option[String], timestamp: String)
-  case class LabelPointSubmission(svImageX: Int, svImageY: Int, canvasX: Int, canvasY: Int, heading: Double, pitch: Double, zoom: Int, canvasHeight: Int, canvasWidth: Int, alphaX: Double, alphaY: Double, lat: Option[Double], lng: Option[Double])
-  case class LabelSubmission(gsvPanoramaId: String, labelType: String, photographerHeading: Double, photographerPitch: Double, deleted: JsBoolean, points: Seq[LabelPointSubmission])
+  case class InteractionSubmission(action: String, gsv_panorama_id: Option[String], lat: Option[Float], lng: Option[Float], heading: Option[Float], pitch: Option[Float], zoom: Option[Int], note: Option[String], timestamp: String)
+  case class LabelPointSubmission(svImageX: Int, svImageY: Int, canvasX: Int, canvasY: Int, heading: Float, pitch: Float, zoom: Int, canvasHeight: Int, canvasWidth: Int, alphaX: Float, alphaY: Float, lat: Option[Float], lng: Option[Float])
+  case class LabelSubmission(gsvPanoramaId: String, labelType: String, photographerHeading: Float, photographerPitch: Float, deleted: JsBoolean, points: Seq[LabelPointSubmission])
   case class TaskSubmission(streetEdgeId: Int, taskStart: String)
   case class AMTAssignmentSubmission(hitId: String, assignmentId: String, assignmentStart: String)
   case class AuditTaskSubmission(assignment: Option[AMTAssignmentSubmission], auditTask: TaskSubmission, labels: Seq[LabelSubmission], interactions: Seq[InteractionSubmission], environment: EnvironmentSubmission)
@@ -30,10 +30,10 @@ object TaskSubmissionFormats {
   implicit val interactionSubmissionReads: Reads[InteractionSubmission] = (
     (JsPath \ "action").read[String] and
       (JsPath \ "gsv_panorama_id").readNullable[String] and
-      (JsPath \ "lat").readNullable[Double] and
-      (JsPath \ "lng").readNullable[Double] and
-      (JsPath \ "heading").readNullable[Double] and
-      (JsPath \ "pitch").readNullable[Double] and
+      (JsPath \ "lat").readNullable[Float] and
+      (JsPath \ "lng").readNullable[Float] and
+      (JsPath \ "heading").readNullable[Float] and
+      (JsPath \ "pitch").readNullable[Float] and
       (JsPath \ "zoom").readNullable[Int] and
       (JsPath \ "note").readNullable[String] and
       (JsPath \ "timestamp").read[String]
@@ -44,22 +44,22 @@ object TaskSubmissionFormats {
       (JsPath \ "sv_image_y").read[Int] and
       (JsPath \ "canvas_x").read[Int] and
       (JsPath \ "canvas_y").read[Int] and
-      (JsPath \ "heading").read[Double] and
-      (JsPath \ "pitch").read[Double] and
+      (JsPath \ "heading").read[Float] and
+      (JsPath \ "pitch").read[Float] and
       (JsPath \ "zoom").read[Int] and
       (JsPath \ "canvas_height").read[Int] and
       (JsPath \ "canvas_width").read[Int] and
-      (JsPath \ "alpha_x").read[Double] and
-      (JsPath \ "alpha_y").read[Double] and
-      (JsPath \ "lat").readNullable[Double] and
-      (JsPath \ "lng").readNullable[Double]
+      (JsPath \ "alpha_x").read[Float] and
+      (JsPath \ "alpha_y").read[Float] and
+      (JsPath \ "lat").readNullable[Float] and
+      (JsPath \ "lng").readNullable[Float]
     )(LabelPointSubmission.apply _)
 
   implicit val labelSubmissionreads: Reads[LabelSubmission] = (
     (JsPath \ "gsv_panorama_id").read[String] and
       (JsPath \ "label_type").read[String] and
-      (JsPath \ "photographer_heading").read[Double] and
-      (JsPath \ "photographer_pitch").read[Double] and
+      (JsPath \ "photographer_heading").read[Float] and
+      (JsPath \ "photographer_pitch").read[Float] and
       (JsPath \ "deleted").read[JsBoolean] and
       (JsPath \ "label_points").read[Seq[LabelPointSubmission]]
     )(LabelSubmission.apply _)
